+++ bundle/src/main/java/com/adobe/acs/commons/httpcache/engine/impl/HttpCacheEngineImpl.java
@@ -400,7 +400,9 @@ public class HttpCacheEngineImpl extends AnnotatedStandardMBean implements HttpC
         // Copy the cached data into the servlet output stream.
         try {
             serveCacheContentIntoResponse(response, cacheContent);
-
+            if (log.isDebugEnabled()) {
+                log.debug("Response delivered from cache for the url [ {} ]", request.getRequestURI());
+            }
             return true;
         } catch (IOException e) {
             throw new HttpCacheDataStreamException("Unable to copy from cached data to the servlet output stream.");
@@ -626,21 +628,12 @@ public class HttpCacheEngineImpl extends AnnotatedStandardMBean implements HttpC
         return tabularData;
     }
 
-    /**
-     * Serves the cache content into the sling response.
-     * Has a fallback for wrappers that do not support the outputstream.
-     * @param response
-     * @param cacheContent
-     * @throws IOException
-     */
     private void serveCacheContentIntoResponse(SlingHttpServletResponse response, CacheContent cacheContent)
-            throws IOException
-    {
-        try{
+            throws IOException {
+        try {
             IOUtils.copy(cacheContent.getInputDataStream(), response.getOutputStream());
-        }catch(IllegalStateException ex) {
-            //for JspResponseServletWrapper and other response wrappers that do not support outputstream.
-            //this will only work for text/html responses.
+        } catch (IllegalStateException ex) {
+            // in this case, either the writer has already been obtained or the response doesn't support getOutputStream()
             IOUtils.copy(cacheContent.getInputDataStream(), response.getWriter(), response.getCharacterEncoding());
         }
     }